- avoid calling strlen() twice in CHUNKPAL.CPP - we find it once, store it in a size_t variable and use that when the length is required - added assert.h include to Chunk.cpp - changed GEI_NOTLOADED define to point to MISSING_TEXTURE rather than '-1' in CHNKTEXI.H - added some spacing to error strings in utilities.cpp to separate the words from the printed LINE numbers etc - changed variable from uint32_t to size_t in TextureManager.cpp as it's used to hold the return from a vector .size() function call, which returns size_t - changed AddItem() in RenderList.cpp to remove reference to flushing as we don't do that yet. Just say the list is full and exit
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@465

+++ 3dc/TextureManager.cpp
@@ -409,7 +409,7 @@ void Tex_ListTextures()
 {
 	std::stringstream ss;
 
-	for (uint32_t i = 0; i < textureList.size(); ++i)
+	for (size_t i = 0; i < textureList.size(); ++i)
 	{
 		if (textureList[i].isValid)
 		{
